requirements-gpu.txt
tensorflow-gpu>=2.0.0
tensorboard
numpy
nltk
tornado
flask
flask-cors